Added singleton and code fixes
