+++ lib/screens/ngGirls.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:ng_poland_conf_next/models/contentful.dart';
-import 'package:ng_poland_conf_next/providers/contentful.dart';
+import 'package:ng_poland_conf_next/providers/simpleContent.dart';
 import 'package:ng_poland_conf_next/widgets/drawer.dart';
 import 'package:provider/provider.dart';
 
-class NgGirls extends StatelessWidget {
+class NgGirls extends StatefulWidget {
   static const routeName = '/NgGirls';
 
   final String title;
@@ -12,21 +12,35 @@ class NgGirls extends StatelessWidget {
   NgGirls({Key key, this.title}) : super(key: key);
 
   @override
+  _NgGirlsState createState() => _NgGirlsState();
+}
+
+class _NgGirlsState extends State<NgGirls> {
+  @override
+  void initState() {
+    Provider.of<SimpleContentProvider>(context, listen: false).fetchData(
+      myId: 'ng-girls-workshops',
+      confId: '2019',
+    );
+    super.initState();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    SimpleContent _simpleContent = Provider.of<ContentfulService>(context)
+    SimpleContent _simpleContent = Provider.of<SimpleContentProvider>(context)
             .simpleContent['ng-girls-workshops'] ??
         null;
 
     return Scaffold(
       appBar: AppBar(
-        title: Text(title),
+        title: Text(widget.title),
         centerTitle: true,
       ),
       drawer: DrawerNg(),
       body: RefreshIndicator(
         onRefresh: () async =>
-            await Provider.of<ContentfulService>(context, listen: false)
-                .getSimpleContentById(
+            await Provider.of<SimpleContentProvider>(context, listen: false)
+                .fetchData(
           myId: 'ng-girls-workshops',
           confId: '2019',
           refresh: true,
