+++ pubspec.yaml
@@ -29,6 +29,7 @@ dependencies:
   shared_preferences: ^0.5.12
   url_launcher: ^5.7.2
   http: ^0.12.2
+  get_it: ^4.0.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
